HSEARCH-3424 Move queries to Spring custom repositories
