+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/integrationtest/showcase/library/repository/DocumentRepository.java
@@ -7,8 +7,9 @@
 package org.hibernate.search.integrationtest.showcase.library.repository;
 
 import org.hibernate.search.integrationtest.showcase.library.model.Document;
+import org.hibernate.search.integrationtest.showcase.library.repository.indexsearch.IndexSearchDocumentRepository;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface DocumentRepository extends CrudRepository<Document, Integer> {
+public interface DocumentRepository extends CrudRepository<Document, Integer>, IndexSearchDocumentRepository {
 }
