+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/integrationtest/showcase/library/repository/LibraryRepository.java
@@ -7,8 +7,9 @@
 package org.hibernate.search.integrationtest.showcase.library.repository;
 
 import org.hibernate.search.integrationtest.showcase.library.model.Library;
+import org.hibernate.search.integrationtest.showcase.library.repository.indexsearch.IndexSearchLibraryRepository;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface LibraryRepository extends CrudRepository<Library, Integer> {
+public interface LibraryRepository extends CrudRepository<Library, Integer>, IndexSearchLibraryRepository {
 }
