+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/integrationtest/showcase/library/repository/PersonRepository.java
@@ -7,8 +7,9 @@
 package org.hibernate.search.integrationtest.showcase.library.repository;
 
 import org.hibernate.search.integrationtest.showcase.library.model.Person;
+import org.hibernate.search.integrationtest.showcase.library.repository.indexsearch.IndexSearchPersonRepository;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person, Integer> {
+public interface PersonRepository extends CrudRepository<Person, Integer>, IndexSearchPersonRepository {
 }
