+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/integrationtest/showcase/library/repository/indexsearch/IndexSearchPersonRepository.java
@@ -1,0 +1,22 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.integrationtest.showcase.library.repository.indexsearch;
+
+import java.util.List;
+
+import org.hibernate.search.integrationtest.showcase.library.model.Person;
+
+public interface IndexSearchPersonRepository {
+
+	List<Person> listTopBorrowers(int offset, int limit);
+
+	List<Person> listTopShortTermBorrowers(int offset, int limit);
+
+	List<Person> listTopLongTermBorrowers(int offset, int limit);
+
+	List<Person> searchPerson(String terms, int offset, int limit);
+}
