+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/integrationtest/showcase/library/service/LibraryService.java
@@ -6,9 +6,7 @@
  */
 package org.hibernate.search.integrationtest.showcase.library.service;
 
-import java.util.Collections;
 import java.util.List;
-import javax.persistence.EntityManager;
 
 import org.hibernate.search.integrationtest.showcase.library.model.Book;
 import org.hibernate.search.integrationtest.showcase.library.model.BookCopy;
@@ -20,8 +18,6 @@ import org.hibernate.search.integrationtest.showcase.library.model.VideoCopy;
 import org.hibernate.search.integrationtest.showcase.library.model.VideoMedium;
 import org.hibernate.search.integrationtest.showcase.library.repository.DocumentCopyRepository;
 import org.hibernate.search.integrationtest.showcase.library.repository.LibraryRepository;
-import org.hibernate.search.mapper.orm.Search;
-import org.hibernate.search.mapper.orm.jpa.FullTextQuery;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -37,9 +33,6 @@ public class LibraryService {
 	@Autowired
 	private DocumentCopyRepository copyRepo;
 
-	@Autowired
-	private EntityManager entityManager;
-
 	public Library create(int id, String name, int collectionSize, double latitude, double longitude, LibraryServiceOption... services) {
 		return libraryRepo.save( new Library( id, name, collectionSize, latitude, longitude, services ) );
 	}
@@ -53,22 +46,6 @@ public class LibraryService {
 	}
 
 	public List<Library> search(String terms, int offset, int limit) {
-		if ( terms == null || terms.isEmpty() ) {
-			return Collections.emptyList();
-		}
-		FullTextQuery<Library> query = Search.getFullTextEntityManager( entityManager )
-				.search( Library.class ).query()
-				.asEntity()
-				.predicate( f -> f.match().onField( "name" ).matching( terms ) )
-				.sort( c -> {
-					c.byField( "collectionSize" ).desc();
-					c.byField( "name_sort" );
-				} )
-				.build();
-
-		query.setFirstResult( offset );
-		query.setMaxResults( limit );
-
-		return query.getResultList();
+		return libraryRepo.search( terms, offset, limit );
 	}
 }
